requirements.txt
Flask==2.3.3
boto3==1.28.85
python-dotenv==1.0.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.11.1
flake8==6.1.0
